refactor SimplifyResults function
