+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "eleventy-plugin-toc",
   "version": "1.0.0",
-  "description": "Generate a TOC for a page rendered with 11ty",
+  "description": "Eleventy filter to generate a Table of Contents in a template",
   "main": ".eleventy.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
