+++ src/NestHeadings.js
@@ -1,35 +1,11 @@
-const simplifyResults = (tag, tags, $) => {
-  const results = []
-
-  $(tag).each((i, el) => {
-    let tag = el.name
-    let id = $(el).attr('id')
-    let text = $(el).text().replace(' #', '')
-    let hierarchy = tags.indexOf(tag)
-    let parent = ( hierarchy > 0 )
-      ? $(el).prevAll(tags[hierarchy - 1]).attr('id')
-      : false
-
-    results.push({
-      order: i,
-      tag,
-      id,
-      text,
-      parent,
-      children: []
-    })
-  })
-
-  return results
-}
-
+const SimplifyResults = require('./SimplifyResults')
 
 const NestHeadings = (tags, $) => {
   let temp = {}
 
   tags
     .forEach((t, i) => {
-      temp[t] = simplifyResults(t, tags, $)
+      temp[t] = SimplifyResults(t, tags, $)
     })
 
   let headings = []
